requirements.txt
asn1crypto==1.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
cssselect==1.1.0
enum34==1.1.10
fuzzywuzzy==0.18.0
idna==3.3
ipaddress==1.0.23
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
nose-parameterized==0.6.0
paramiko==2.8.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyquery==1.4.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
